# docs/swagger.yaml
basePath: /
definitions:
  dto.AppointmentResponse:
    properties:
      created_at:
        type: string
      doctor_id:
        type: string
      duration:
        type: integer
      id:
        type: string
      notes:
        type: string
      patient_id:
        type: string
      reason:
        type: string
      scheduled_at:
        type: string
      service_id:
        type: string
      service_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateAppointmentRequest:
    properties:
      appointment_date:
        example: "2025-11-15"
        type: string
      appointment_time:
        example: "10:00"
        type: string
      doctor_id:
        example: uuid
        type: string
      reason:
        example: Consulta general
        type: string
      service_id:
        example: uuid
        type: string
    type: object
...
  dto.DashboardSummary:
    properties:
      cancellation_rate:
        type: number
      cancelled_appointments:
        type: integer
      completed_appointments:
        type: integer
      confirmed_appointments:
        type: integer
      pending_appointments:
        type: integer
      total_appointments:
        type: integer
      total_doctors:
        type: integer
      total_patients:
        type: integer
      total_revenue:
        type: number
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  dto.RevenueByService:
    properties:
      revenue:
        type: number
      service_id:
        type: string
      service_name:
        type: string
      total_citas:
        type: integer
    type: object
...
  dto.TopDoctor:
    properties:
      completed_appointments:
        type: integer
      doctor_id:
        type: string
      doctor_name:
        type: string
      total_appointments:
        type: integer
    type: object
  dto.TopService:
    properties:
      service_id:
        type: string
      service_name:
        type: string
      total_citas:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: zensoftsas@gmail.com
    name: Zensoft Team
    url: https://github.com/zensoftsas/sistema-reservas
  description: |
    API completa para gestión de clínica médica con sistema de citas, servicios, horarios personalizados y analytics.

    Características principales:
    - Autenticación JWT con roles (admin, doctor, patient)
    - Sistema de citas médicas con validaciones
    - Servicios médicos con precios y duraciones
    - Horarios personalizados por doctor
    - Cálculo dinámico de slots disponibles
    - Analytics y reportes para administradores
    - Notificaciones por email (SendGrid)
    - Recordatorios automáticos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sistema de Reservas - Clínica Internacional API
  version: "1.0"
paths:
  /api/analytics/dashboard:
    get:
      consumes:
      - application/json
      description: Retorna resumen con KPIs principales del sistema (solo admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dashboard summary
      tags:
      - Analytics
  /api/analytics/revenue:
    get:
      consumes:
      - application/json
      description: Retorna ingresos agrupados por servicio (solo admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RevenueByService'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Estadísticas de ingresos
      tags:
      - Analytics
  /api/analytics/top-doctors:
    get:
      consumes:
      - application/json
      description: Retorna ranking de doctores por cantidad de citas (solo admin)
      parameters:
      - default: 10
        description: Número de doctores a retornar
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TopDoctor'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top doctores más solicitados
      tags:
      - Analytics
  /api/analytics/top-services:
    get:
      consumes:
      - application/json
      description: Retorna ranking de servicios por demanda (solo admin)
      parameters:
      - default: 10
        description: Número de servicios a retornar
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TopService'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Top servicios más populares
      tags:
      - Analytics
  /api/appointments:
    post:
      consumes:
      - application/json
      description: Crear una nueva cita médica con validaciones de disponibilidad
      parameters:
      - description: Datos de la cita
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear cita médica
      tags:
      - Appointments
  /api/appointments/my:
    get:
      consumes:
      - application/json
      description: Retorna todas las citas del usuario autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AppointmentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener mis citas
      tags:
      - Appointments
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...